Fix up frozen indices split package
Moves the frozen engine and friends to `o.e.i.e.frozen`.
